requirements.txt
Django==3.1.3
django-admin-interface==0.14.1
Pillow==8.0.1
django-crispy-forms
djangorestframework
markdown
django-filter
social-auth-app-django